reference/requirements.txt
requests
tensorflow==2.6
keras==2.6.0
pyyaml
scipy
easydict
boto3
opencv-python==4.6.0.66
mediapipe
neurokit2==0.2.1
python-dotenv
requests
kapre==0.3.7
